Extract check_credentials from __init__
