+++ tests/unit/cmk/update_config/plugins/actions/test_password_hashes.py
@@ -7,7 +7,7 @@ import datetime
 
 import pytest
 
-from tests.unit.cmk.gui.test_userdb import _load_users_uncached
+from tests.unit.cmk.gui.userdb.test_userdb import _load_users_uncached
 
 from cmk.utils.crypto.password import PasswordHash
 from cmk.utils.crypto.password_hashing import is_unsupported_legacy_hash
